+++ order-service/src/main/java/com/bernarsk/orderservice/model/OrderDetails.java
@@ -13,8 +13,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -24,14 +22,14 @@ import java.util.List;
 public class OrderDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
     private Order order;
 
     @Column
-    private Integer productId;
+    private Long productId;
 
     @Column
     private Integer quantity;
