+++ order-service/src/main/java/com/bernarsk/orderservice/repository/OrderDetailsRepository.java
@@ -3,6 +3,6 @@ package com.bernarsk.orderservice.repository;
 import com.bernarsk.orderservice.model.OrderDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
 
 }
